move endpoints to lib root
