+++ lib/endpoints/apps.rb
